pyproject.toml
[project]
name = "networds"
version = "0.1.0"
description = ""
authors = [
    {name = "Akshay-Shingala",email = "akshay.shingala@yudiz.com"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "sqlalchemy (>=2.0.42,<3.0.0)",
    "mongoconnect (>=0.0.6,<0.0.7)",
    "strawberry-graphql[fastapi] (>=0.278.0,<0.279.0)",
    "databases[sqlite] (>=0.9.0,<0.10.0)",
    "graphene (>=3.4.3,<4.0.0)",
    "django (>=5.2.4,<6.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
black = "^25.1.0"
isort = "^6.0.1"
flake8 = "^7.3.0"

[tool.poetry]
package-mode = false
